step2_preprocessing/requirements.txt
# Core dependencies from Step 1
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.1

# Deep learning and transformers
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Clinical NLP and text processing
scispacy>=0.5.3
spacy>=3.7.0
# For scispacy models: python -m spacy download en_core_sci_md
medcat>=1.9.0  # Optional: Medical Concept Annotation Tool

# LangChain and LLM integrations
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.8.0

# Embeddings and similarity
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # Or faiss-gpu for GPU acceleration

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
